requirements.txt
streamlit
langchain
langgraph
openai
whisper
sounddevice
scipy
python-dotenv
langchain_openai
transformers
torch
sounddevice
numpy
gtts
